Change virtiofs service's Name and DisplayName
Fixes #23

Previously we were essentially doing:

  sc create "Virtio FS Service" ... DisplayName=VirtioFsSvc

but now we're essentially doing:

  sc create VirtioFsSvc ... DisplayName="Virtio FS Service"

The previous behavior was confusing because it made all future
interacting with the service (using tools such as `sc`) look like the
service's DisplayName was being specified where the command actually
expects (requires!) the service's Name to be specified.

Note that this commit uses "VirtioFsSvc" as the service Name instead of
"VirtioFSService". That's for consistency with the instructions that
people have been using when following:

  https://virtio-fs.gitlab.io/howto-windows.html

This commit also adds a "Description" field to help people determine
what the service is for when they encounter it in the Services app, for
example.

Signed-off-by: Jonathan Watt <jwatt@jwatt.org>